+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 - [server] Restart count on 'app info' output
+- [client] Validate 'deploy' command parameters
 
 ## [0.6.0] - 2017-08-23
 ### Added